Use Adapt.jl to change storage and element type
In order to eventually support GPU computation we need
to use Adapt.jl to allow GPU backend packages to swap
out host-array types like `CuArray` with device-side types
like `CuDeviceArray`.

Additionally this will allow us to change the element type
of a simulation by using `adapt(Array{Float32}`.


restore elixir

offload compute_coefficients

fmt

test native version as well

adapt 1D and 3D version

Downgrade compat with Adapt

update requires to 1.3

add support for AMDGPU

fix doctest

Use `u_ode` to determine the computational backend


Use KA 0.9.31

handle VectorOfArray in trixi_backend

fixup: runtests

format

fix trixi_backend for RecursiveArrayTools{StaticArray}

fixup: amdgpu test

use unsafe_wrap with lock=false for AMDGPU

Update src/solvers/dg.jl

Update ext/TrixiAMDGPUExt.jl


Update src/auxiliary/containers.jl


use eleixi_advection_basic_gpu.jl for adapt test

Update examples/p4est_2d_dgsem/elixir_advection_basic_gpu.jl


something on KA

move to RecursiveArrayTools:VectorOfArray

fix stuff related to VectorOfArray

fix parallel TreeMesh

more fixes

StaticArrays

ADTypes?

even more

more

more

doubled during merge

+++ src/semidiscretization/semidiscretization.jl
@@ -103,7 +103,7 @@ function semidiscretize(semi::AbstractSemidiscretization, tspan;
             storage_type = Array
         end
         if real_type === nothing
-            real_type = real(semi)
+            real_type = Float64
         end
         semi = trixi_adapt(storage_type, real_type, semi)
         if eltype(tspan) !== real_type
